comment dir rename discussion
