+++ dashboard/src/components/discussion/AnchorCard.tsx
