+++ dashboard/src/components/template/Wbw/WbwPali.tsx
@@ -12,7 +12,7 @@ import WbwDetail from "./WbwDetail";
 import { IWbw, TWbwDisplayMode } from "./WbwWord";
 import { bookMarkColor } from "./WbwDetailBookMark";
 import WbwVideoButton from "./WbwVideoButton";
-import CommentBox from "../../comment/CommentBox";
+import CommentBox from "../../discussion/CommentBox";
 import PaliText from "./PaliText";
 import store from "../../../store";
 import { lookup } from "../../../reducers/command";
